Delay generating source dummy k-mers until after k-mers are sorted and merged (#157)
Instead of generating dummy k-mers while tokenizing the input sequences. the dummy source k-mers are generating during the final merge.
For in-memory merge, the dummy-1 kmers are generated and checked for redundancy at the same time, see comments in code.
For disk-based merge, the dummy-1 kmers are generated in N different files, one file for each character in the alphabet. Since the generated files are sorted, all we need to do is to merge the N files, which removes the need of using SortedSetDisk and ChunkedWaitQueue.

Another simplification that I made was to not store counts for dummy k-mers in the intermediate files.

Co-authored-by: Daniel Danciu <dd@daedalean.ai>
Co-authored-by: Mikhail Karasikov <mike-karas@ya.ru>